Variation 2
In this variation 2, the flow field should react to the mouse position, interacting with it. Started with adding the mouse X and Y parameters to the line function, and checking through the if statement if it is on the mouse radius something should happen otherwise keeps the same flow field.